chore(ts-utils): address review feedback on Result.shouldNotFail
- Extract _IShouldNotFailFrame, _parseStackFrame, _findShouldNotFailFrame,
  _normalizeShouldNotFailFnName, and _formatShouldNotFailMessage into
  packlets/base/shouldNotFail.ts so result.ts stays focused on the Result
  hierarchy itself (per ErikFortune review).
- Update _findShouldNotFailFrame JSDoc to accurately describe the filter:
  parsed function name only — never the raw line text (per copilot review;
  the earlier "source line" wording risked reintroducing the file-path
  collision bug).
- Remove ambiguous {@link IResult.orThrow | orThrow} link from the
  shouldNotFail JSDoc — orThrow is overloaded and the link triggered an
  ae-unresolved-link warning in etc/ts-utils.api.md (per copilot review).
- Replace `e as Error` unsafe casts in shouldNotFail.test.ts with the
  existing _errorMessage helper, and replace the `Error as { ... }` cast
  on captureStackTrace mutation with Reflect.set (per copilot review).
